Make `create!` more permissive and more performant (#30)
create! used to instantiate a whole entity from the attrs in order to
build a url chunk that didn't depend on the entity at all. This was a
waste of time and memory and made it so that we couldn't pass through
attrs with modified keys (like those with odata.bind annotations) unless
we also rebuilt entities to be aware of annotations. Also remove parens
from returned ids.