[DWF-2442] Make nav prop and ref constraint inheritance work (#39)
Previously, if a entity inherited from another type, the child would
know about the properties of the parent type, but the navigation
properties and referential constraints did not reflect those of the
parent type. Now if one calls either navigation_properties or
referential_constraints_for_entity, the result will include properties
for all parent types as well as the child.